requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
environ==1.0
Django==2.2.10
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.10.3
environ==1.0
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
Markdown==3.1.1
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6